requirements.txt
cffi~=1.16.0
numpy~=1.26.4
pycparser~=2.21
pygame-ce~=2.4.1
pymunk~=6.6.0
PyYAML~=6.0.1
